Update loglines test  - updated filename to match tooling  - import functions, not module  - use testcase.assertEqual instead of assert
